+++ spec/gorillib/builder_spec.rb
@@ -3,6 +3,7 @@ require 'spec_helper'
 # libs under test
 require 'gorillib/builder'
 require 'gorillib/builder/field'
+require 'gorillib/collection/model_collection'
 
 # testing helpers
 require 'gorillib/hash/compact'
@@ -116,10 +117,10 @@ describe Gorillib::Builder, :model_spec => true, :builder_spec => true do
 
   context 'collections' do
     subject{ garage }
-    let(:sample_val){ Gorillib::Collection.receive([wildcat], car_class, :name) }
+    let(:sample_val){ Gorillib::ModelCollection.receive([wildcat], car_class, :name) }
     let(:raw_val   ){ [ wildcat.attributes ] }
     it_behaves_like "a model field", :cars
-    it("#read_attribute is an empty collection if never set"){ subject.read_attribute(:cars).should == Gorillib::Collection.new }
+    it("#read_attribute is an empty collection if never set"){ subject.read_attribute(:cars).should == Gorillib::ModelCollection.new }
 
     it 'a collection holds named objects' do
       garage.cars.should be_empty
@@ -140,12 +141,13 @@ describe Gorillib::Builder, :model_spec => true, :builder_spec => true do
 
       # examine the whole collection
       garage.cars.keys.should == [:cadzilla, :wildcat, :ford_39]
-      garage.cars.should == Gorillib::Collection.receive([cadzilla, wildcat, ford_39], car_class, :name)
+      garage.cars.should == Gorillib::ModelCollection.receive([cadzilla, wildcat, ford_39], car_class, :name)
     end
+
     it 'lazily autovivifies collection items' do
       garage.cars.should be_empty
       garage.car(:chimera).should be_a(car_class)
-      garage.cars.should == Gorillib::Collection.receive([{:name => :chimera}], car_class, :name)
+      garage.cars.should == Gorillib::ModelCollection.receive([{:name => :chimera}], car_class, :name)
     end
 
     context 'collection getset method' do